Add QueryLoader to test bench

+++ src/MigrationAssistant.Shared/Records/QueryRecord.cs
@@ -4,6 +4,6 @@ public class QueryRecord
 {
     public Guid TestsetId { get; set; }
     public Guid QueryId { get; set; }
-    public int SessionId { get; set; }
+    public UInt16 SessionId { get; set; }
     public string QueryText { get; set; } 
 }
